Create Listings and Single Listings page

+++ src/app/api/(property)/removeproperties/route.js
@@ -1,5 +1,8 @@
+/** @format */
+
 import Property from "@/models/property.model";
 import { connect } from "@/utils/dbConnect";
+import { DeleteOnCloudinary } from "@/utils/deleteOnCloudinary";
 import { NextResponse } from "next/server";
 connect();
 
@@ -9,16 +12,29 @@ export const POST = async (request) => {
     const { id } = reqBody;
 
     if (!id) {
-      NextResponse.json(
+      return NextResponse.json(
         { error: "To remove property there must be specific id" },
         { status: 400 },
       );
     }
+    const property = await Property.findById(id);
 
+    if (!property) {
+      return NextResponse.json(
+        { error: "Failed! Try again later" },
+        { status: 400 },
+      );
+    }
+
+    try {
+      await DeleteOnCloudinary(property.imagePublicId);
+    } catch (error) {
+      throw new Error("Cloudinary Error :" + error.message);
+    }
     const deletedProperty = await Property.findByIdAndDelete(id);
 
     if (!deletedProperty) {
-      NextResponse.json(
+      return NextResponse.json(
         { error: "Error in Deleting Property" },
         { status: 400 },
       );
